fixing issues...
